up_lzy
1. changes the places where lzy child need to updated with macro 'mch' (mark child)
2. arranged the code, and added printBoth() function.
NOTE:- In the build , the commented code are for the case when inputting of data into the t[] is not done manually in the main() funciton.
NOTE2: The push_up(..) is the place where, the parent is changes in accrodance with the node children. The relationship between parent and children is governed by the QUESTION.
NOTE3: In the updateRANGE_LAZY(..) the case where node needs to be updated when they are marked as lazy, and case when the update in within range. The implementation of these two, depends HEAVILY UPON QUESTION.
NOTE4: Similary for queryLAZY(..) the case when the node is marked as lazy depends HEAVILY UPON QUESTION. And this one is exactly same as its counterpart in updateRANGE_LAZY(..)